+++ superwall/src/main/java/com/superwall/sdk/dependencies/DependencyContainer.kt
@@ -299,6 +299,7 @@ class DependencyContainer(
             PaywallMessageHandler(
                 sessionEventsManager = sessionEventsManager,
                 factory = this@DependencyContainer,
+                ioScope = ioScope,
             )
 
         val paywallView =